Rename `rich-formatting` extension to `text-styling`
The only remaining functionality there is applying
style marks to text.
