+++ packages/slate-editor/src/extensions/text-styling/TextStylingExtension.tsx
@@ -1,0 +1,20 @@
+import type { Extension } from '@prezly/slate-commons';
+import type { KeyboardEvent } from 'react';
+import type { Editor } from 'slate';
+
+import { Text } from './components';
+import { createDeserialize } from './createDeserialize';
+import * as OnKeyDown from './onKeyDown';
+
+export const EXTENSION_ID = 'TextStylingExtension';
+
+export function TextStylingExtension(): Extension {
+    return {
+        id: EXTENSION_ID,
+        deserialize: createDeserialize(),
+        onKeyDown: (event: KeyboardEvent, editor: Editor) => {
+            OnKeyDown.onHotkeyDoMarks(event, editor);
+        },
+        renderLeaf: Text,
+    };
+}
