+++ packages/slate-editor/src/extensions/text-styling/index.ts
@@ -1,4 +1,4 @@
-export { RichFormattingExtension, EXTENSION_ID } from './RichFormattingExtension';
+export { TextStylingExtension, EXTENSION_ID } from './TextStylingExtension';
 
 export { Text } from './components';
 export type { RichFormattingExtensionParameters, RichTextElementType } from './types';
