+++ packages/slate-editor/src/extensions/text-styling/lib/detectMarks/detectMark.ts
