+++ packages/slate-editor/src/extensions/text-styling/lib/detectMarks/isItalic.ts
