+++ packages/slate-editor/src/modules/rich-formatting-menu/types.ts
@@ -1,5 +1,5 @@
 import type { PARAGRAPH_NODE_TYPE } from '@prezly/slate-types';
 
-import type { ElementType } from '#extensions/rich-formatting';
+import type { ElementType } from '#extensions/text-styling';
 
 export type Formatting = typeof PARAGRAPH_NODE_TYPE | ElementType | 'multiple';
